Re-arrange the code to group it in more module/entity-specific way
