+++ rule-keeper/tag_suggesters.py
@@ -1,6 +1,27 @@
-from post_data_extractor import PostData
+from post import PostData, PostDataExtractor
 from rule_keeper import RuleCheckResults
 import jellyfish
+import os
+
+
+def find_existing_tags(
+        data_extractor: PostDataExtractor,
+        posts_source_directory: str,
+        post_files_to_omit: list[str]
+) -> list[str]:
+    filenames_to_omit = [os.path.basename(post_file_path) for post_file_path in post_files_to_omit]
+
+    existing_tags = []
+
+    for filename in os.listdir(posts_source_directory):
+        if filename in filenames_to_omit:
+            continue
+
+        post_data = data_extractor.extract_data(os.path.join(posts_source_directory, filename))
+        if 'tags' in post_data.metadata:
+            existing_tags = existing_tags + post_data.metadata['tags']
+
+    return existing_tags
 
 
 class ExistingTagsSuggester:
